build(android): extract shared android-app-signing convention (#5072)
control-proxy/build.gradle.kts and playground/app/build.gradle.kts carried a
byte-identical ~35-line release-signing block (keystore env/property reads,
signingConfigs { create("release") { … } }, and the
signingConfig = if (hasReleaseSigning) release else debug assignments on both
build types). Extract it into an automobile.android-app-signing precompiled
convention in android/build-logic, applied via
id("automobile.android-app-signing") in the two application modules.

Keystore path resolution moves from rootProject.file(path) to
rootDir.resolve(path) (Isolated-Projects safe; matches the existing
automobile.detekt convention's rootDir usage). isMinifyEnabled/proguardFiles
stay in the modules — no behavior change there.

Validated:
- :control-proxy:assembleRelease + :playground:app:assembleRelease with no
  keystore → both fall back to the debug cert (CN=AutoMobile).
- Same with RELEASE_KEYSTORE_* set → both sign with the release cert
  (verified via apksigner verify -v --print-certs).
- IP dry-run (-Dorg.gradle.unsafe.isolated-projects=true) reports zero
  config-phase problems for both modules.

Closes #5072
