+++ android/control-proxy/build.gradle.kts
@@ -1,29 +1,10 @@
 plugins {
   id("automobile.kotlin-common")
+  id("automobile.android-app-signing")
   alias(libs.plugins.android.application)
   alias(libs.plugins.kotlin.serialization)
   alias(libs.plugins.compose.compiler)
 }
-
-// Signing configuration: reads from environment variables (CI) or gradle.properties (local)
-// Paths are resolved relative to project root using rootProject.file()
-val releaseStoreFilePath: String? =
-  System.getenv("RELEASE_KEYSTORE_PATH") ?: findProperty("RELEASE_KEYSTORE_PATH") as String?
-val releaseStorePassword: String? =
-  System.getenv("RELEASE_KEYSTORE_PASSWORD") ?: findProperty("RELEASE_KEYSTORE_PASSWORD") as String?
-val releaseKeyAlias: String? =
-  System.getenv("RELEASE_KEY_ALIAS") ?: findProperty("RELEASE_KEY_ALIAS") as String?
-val releaseKeyPassword: String? =
-  System.getenv("RELEASE_KEY_PASSWORD") ?: findProperty("RELEASE_KEY_PASSWORD") as String?
-val releaseStoreFile: File? = releaseStoreFilePath?.let { path ->
-  val file = File(path)
-  if (file.isAbsolute) file else rootProject.file(path)
-}
-val hasReleaseSigning =
-  releaseStoreFile?.exists() == true &&
-    !releaseStorePassword.isNullOrBlank() &&
-    !releaseKeyAlias.isNullOrBlank() &&
-    !releaseKeyPassword.isNullOrBlank()
 
 android {
   namespace = "dev.jasonpearson.automobile.ctrlproxy"
@@ -40,33 +21,10 @@ android {
     testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
   }
 
-  signingConfigs {
-    create("release") {
-      storeFile = releaseStoreFile
-      storePassword = releaseStorePassword
-      keyAlias = releaseKeyAlias
-      keyPassword = releaseKeyPassword
-    }
-  }
-
   buildTypes {
     release {
       isMinifyEnabled = false
       proguardFiles(getDefaultProguardFile("proguard-android-optimize.txt"), "proguard-rules.pro")
-      signingConfig =
-        if (hasReleaseSigning) {
-          signingConfigs.getByName("release")
-        } else {
-          signingConfigs.getByName("debug")
-        }
-    }
-    debug {
-      signingConfig =
-        if (hasReleaseSigning) {
-          signingConfigs.getByName("release")
-        } else {
-          signingConfigs.getByName("debug")
-        }
     }
   }
